linux-generic: random: split from crypto module
While random and crypto might share some implementation details, in case
of linux-generic and linux-DPDK it might be easier to split them to two
different files.
