+++ platform/linux-generic/Makefile.am
@@ -175,6 +175,7 @@ __LIB__libodp_linux_la_SOURCES = \
 			   odp_queue_if.c \
 			   odp_queue_lf.c \
 			   odp_queue_scalable.c \
+			   odp_random.c \
 			   odp_rwlock.c \
 			   odp_rwlock_recursive.c \
 			   odp_schedule_basic.c \
